% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AnnoBroadGseaRes-class}
\alias{AnnoBroadGseaRes-class}
\title{Annotated BROAD GSEA Results for one contrast}
\value{
An object of class \code{AnnoBroadGseaRes}.
}
\description{
Annotated BROAD GSEA Results for one contrast
}
