% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{AnnoBroadGseaRes}
\alias{AnnoBroadGseaRes}
\title{Convert a list of AnnoBroadGseaResItem objects to a list}
\usage{
AnnoBroadGseaRes(object)
}
\arguments{
\item{object}{A list of AnnoBroadGseaResItem}
}
\value{
An \code{AnnoBroadGseaRes} object
}
\description{
Convert a list of AnnoBroadGseaResItem objects to a list
}
