% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AnnoBroadGseaResItem-class}
\alias{AnnoBroadGseaResItem-class}
\title{Annotated BROAD GSEA result item}
\description{
Annotated BROAD GSEA result item
}
\section{Slots}{

\describe{
\item{\code{gsGenes}}{Vector of character strings, gene-set genes}

\item{\code{gsGeneValues}}{Vector of numeric values, statistics of gene-set genes}
}}

