% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoBroadGseaResItem.R
\name{annoBroadGseaResItem}
\alias{annoBroadGseaResItem}
\title{Convert a broadGseaResItem object to an annoBroadGseaResItem object}
\usage{
annoBroadGseaResItem(object, genes, geneValues)
}
\arguments{
\item{object}{A broadGseaResItem object}

\item{genes}{A character string vector}

\item{geneValues}{A numeric vector}
}
\value{
An annobroadGseaResItem object
}
\description{
Convert a broadGseaResItem object to an annoBroadGseaResItem object
}
