% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{AnnoBroadGseaResList-class}
\alias{AnnoBroadGseaResList-class}
\title{A list of AnnoBroadGseaRes objects}
\value{
An object of class \code{AnnoBroadGseaResList}.
}
\description{
A list of AnnoBroadGseaRes objects
}
