% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{appendGeneSets}
\alias{appendGeneSets}
\title{Append one or more instances of GeneSets to an existing instance}
\usage{
appendGeneSets(geneSets, newGeneSets, ...)
}
\arguments{
\item{geneSets}{An existing instance of GeneSets to which other instances should be appended}

\item{newGeneSets}{An instance of GeneSets that will be appended to \code{geneSets}}

\item{...}{Other instance(s) of GeneSets that will be appended consequentially}
}
\description{
Append one or more instances of GeneSets to an existing instance
}
\examples{
gmtFile <- system.file("extdata", "example.gmt", package="ribiosGSEA")
mySet <- readGmt(gmtFile)
myFakeSet <- readGmt(CategoryA=gmtFile, CategoryB=gmtFile)
anotherFakeSet <- readGmt(gmtFile, gmtFile, category=c("CategoryC", "CategoryD"))
mySetAppended <- appendGeneSets(mySet, myFakeSet)
mySetAppendedTwice <- appendGeneSets(mySet, myFakeSet, anotherFakeSet
}
