% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{as.data.frame,FisherResultList-method}
\alias{as.data.frame,FisherResultList-method}
\title{Convert an FisherResultList object into a data.frame}
\usage{
\S4method{as.data.frame}{FisherResultList}(x, row.names = NULL)
}
\arguments{
\item{x}{An FisherResultList object}

\item{row.names}{Character strings.}
}
\value{
A \code{data.frame}
}
\description{
Convert an FisherResultList object into a data.frame
}
