% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{as.matrix.TwoGroupExprsSimulator}
\alias{as.matrix.TwoGroupExprsSimulator}
\title{Return matrix from TwoGroupExprsSimulator}
\usage{
\method{as.matrix}{TwoGroupExprsSimulator}(x, ...)
}
\value{
A matrix
}
\description{
Return matrix from TwoGroupExprsSimulator
}
