% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{broadGseaResItem-class}
\alias{broadGseaResItem-class}
\title{A S4 class representing the atom structure of results of the BROAD GSEA tool}
\description{
A S4 class representing the atom structure of results of the BROAD GSEA tool
}
\section{Slots}{

\describe{
\item{\code{geneSet}}{Character, gene-set name}

\item{\code{es}}{Numeric, enrichment score}

\item{\code{nes}}{Numeric, normalised enrichment score}

\item{\code{np}}{Numeric}

\item{\code{fdr}}{Numeric, false discovery rate}

\item{\code{FWER}}{Numeric, family-wise error rate}

\item{\code{geneIndices}}{Integer vector, gene indices}

\item{\code{esProfile}}{Numeric, enrichment score profile}

\item{\code{coreEnrichThr}}{Numeric}
}}

