% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBroadGSEAcomm.R
\name{buildBroadGSEAcomm}
\alias{buildBroadGSEAcomm}
\title{Build the command-line command to run BROAD GSEA}
\usage{
buildBroadGSEAcomm(
  gseaJar,
  javaBin,
  rnkFiles,
  gmtFile,
  chipFile,
  nperm = 1000L,
  collapse = FALSE,
  plotTopX = 25,
  outdir = "./",
  addShebang = TRUE
)
}
\arguments{
\item{gseaJar}{Character string, full file name of BROAD GSEA (gene
permutation) jar file}

\item{javaBin}{Character string, java binary file}

\item{rnkFiles}{Character string, rank files}

\item{gmtFile}{A \code{GMT} file encoding GMT files to be used}

\item{chipFile}{A \code{CHIP} file encoding feature annotation}

\item{nperm}{Integer, number of permutations}

\item{collapse}{Logical, whether to collapse duplicated features}

\item{plotTopX}{Integer, top gene-sets to be visualized}

\item{outdir}{Character string, the path of output}

\item{addShebang}{Logical, whether to add Shebang to the script

The command builds command-line command to run gene-permutation GSEA
over many rank files.}
}
\description{
Build the command-line command to run BROAD GSEA
}
