% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{camera.LimmaVoomResult}
\alias{camera.LimmaVoomResult}
\title{Run the CAMERA method using LimmaVoomResult}
\usage{
\method{camera}{LimmaVoomResult}(y, gmtList, doParallel = FALSE)
}
\arguments{
\item{y}{A LimmaVoomResult object}

\item{gmtList}{Gene set collections, for example read by
\code{\link[BioQC]{readGmt}}}

\item{doParallel}{Logical, whether \code{parallel::mclapply} should be used. Since at the current setting it makes a job running forever, use \code{TRUE} only if you are debugging the code.

Note that the LimmaVoomResult object must have a column 'GeneSymbol' in its
\code{fData}.}
}
\description{
Run the CAMERA method using LimmaVoomResult
}
\examples{
exMat <- matrix(rpois(120, 10), nrow=20, ncol=6)
exGroups <- gl(2,3, labels=c("Group1", "Group2"))
exDesign <- model.matrix(~0+exGroups)
colnames(exDesign) <- levels(exGroups)
exContrast <- matrix(c(-1,1), ncol=1, dimnames=list(c("Group1", "Group2"), c("Group2.vs.Group1")))
exDescon <- DesignContrast(exDesign, exContrast, groups=exGroups)
exFdata <- data.frame(GeneSymbol=sprintf("Gene\%d", 1:nrow(exMat)))
exPdata <- data.frame(Name=sprintf("Sample\%d", 1:ncol(exMat)),
                     Group=exGroups)
exDgeList <- DGEList(exMat, genes=exFdata, samples=exPdata)
exDgeList <- edgeR::estimateDisp(exDgeList, exDesign)
edgeObj <- EdgeObject(exDgeList, exDescon)
limmaVoomRes <- dgeWithLimmaVoom(edgeObj)
exGmt <- GmtList(list(GeneSet1=sprintf("GeneSymbol\%d", 1:5),
  GeneSet2=sprintf("GeneSymbol\%d", 6:10)))
  
camera(limmaVoomRes, exGmt)
}
