% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{cameraLimmaVoomResultsByContrast}
\alias{cameraLimmaVoomResultsByContrast}
\title{Apply the CAMERA method to a DGEList object}
\usage{
cameraLimmaVoomResultsByContrast(
  limmaVoomResults,
  index,
  doParallel = FALSE,
  ...
)
}
\arguments{
\item{limmaVoomResults}{A LimmaVoomResults object, with GeneSymbol available}

\item{index}{List of integer indices of genesets, names are names of gene
sets}

\item{doParallel}{Logical, whether \code{parallel::mclapply} should be used. Since at the current setting it makes a job running forever, use \code{TRUE} only if you are debugging the code.}

\item{...}{Not used}
}
\value{
A \code{data.frame} containing CAMERA results.
}
\description{
Apply the CAMERA method to a DGEList object
}
\examples{
exMat <- matrix(rpois(120, 10), nrow=20, ncol=6)
exGroups <- gl(2,3, labels=c("Group1", "Group2"))
exDesign <- model.matrix(~0+exGroups)
colnames(exDesign) <- levels(exGroups)
exContrast <- matrix(c(-1,1), ncol=1, dimnames=list(c("Group1", "Group2"), c("Group2.vs.Group1")))
exDescon <- DesignContrast(exDesign, exContrast, groups=exGroups)
exFdata <- data.frame(GeneSymbol=sprintf("Gene\%d", 1:nrow(exMat)))
exPdata <- data.frame(Name=sprintf("Sample\%d", 1:ncol(exMat)),
                     Group=exGroups)
exDgeList <- DGEList(exMat, genes=exFdata, samples=exPdata)
exDgeList <- edgeR::estimateDisp(exDgeList, exDesign)
edgeObj <- EdgeObject(exDgeList, exDescon)
limmaVoomRes <- ribiosNGS::dgeWithLimmaVoom(edgeObj)
cameraLimmaVoomResultsByContrast(limmaVoomRes, index=c(1:5))
cameraLimmaVoomResultsByContrast(limmaVoomRes, index=list(GS1=1:5, GS2=6:10))

}
