% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{cameraTable2graph}
\alias{cameraTable2graph}
\title{Convert a CAMERA table into a graph}
\usage{
cameraTable2graph(df, jacThr = 0.25, plot = TRUE, ...)
}
\arguments{
\item{df}{Data.frame, CAMERA results}

\item{jacThr}{Numeric, between 0 and 1, Jaccard Index threshold}

\item{plot}{Logical, whether plotting the results}

\item{...}{Passed to \code{plot}}
}
\description{
Convert a CAMERA table into a graph
}
