% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{contrastNames,EdgeGSE-method}
\alias{contrastNames,EdgeGSE-method}
\title{Extract contrast names from an EdgeGSE object}
\usage{
\S4method{contrastNames}{EdgeGSE}(object)
}
\arguments{
\item{object}{An \code{EdgeGSE} object}
}
\description{
Extract contrast names from an EdgeGSE object
}
