% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{EdgeGSE-class}
\alias{EdgeGSE-class}
\title{An EdgeGSE object contains gene-sets, enrichment method, and 
enrichment tables besides EdgeResult}
\description{
An EdgeGSE object contains gene-sets, enrichment method, and 
enrichment tables besides EdgeResult
}
\section{Slots}{

\describe{
\item{\code{geneSets}}{A GmtList}

\item{\code{method}}{Gene-set enrichment method}

\item{\code{enrichTables}}{A data.frame}
}}

