% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{EdgeGSE}
\alias{EdgeGSE}
\title{Build an EdgeGSE object}
\usage{
EdgeGSE(edgeObj, gmtList)
}
\arguments{
\item{edgeObj}{An \code{EdgeObject}}

\item{gmtList}{A \code{GmtList} object}
}
\value{
An EdgeGSE object, with enrichTables as \code{NULL}
}
\description{
Build an EdgeGSE object
}
