% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{expandCameraTableGenes}
\alias{expandCameraTableGenes}
\title{Expand genes in the CAMERA result table}
\usage{
expandCameraTableGenes(tbl)
}
\arguments{
\item{tbl}{A \code{data.frame}}
}
\value{
A longer \code{data.frame}, with each row one gene.
}
\description{
Expand genes in the CAMERA result table
}
