% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strWithNumber.R
\name{factorByNumberInStr}
\alias{factorByNumberInStr}
\title{Make a factor vector from a character vector by the order of the parsed numbers}
\usage{
factorByNumberInStr(str, decreasing = TRUE)
}
\arguments{
\item{str}{Strings}

\item{decreasing}{Logical, whether decreasing or increasing order is desied, passed to \code{order}.}
}
\description{
Make a factor vector from a character vector by the order of the parsed numbers
}
\examples{
factorByNumberInStr(c("D1", "D10", "D15", "D3.5"))
factorByNumberInStr(c("D1", "D10", "D15", "D3.5"), decreasing=FALSE)
}
\seealso{
\code{\link{orderByNumberInStr}}, which returns the order of strings
by numbers in them
}
