% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fdrValue}
\alias{fdrValue}
\title{Return FDR values}
\usage{
fdrValue(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{
A numeric vector of FDR values.
}
\description{
Return FDR values
}
