% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{filterBySize}
\alias{filterBySize}
\title{Filter by size}
\usage{
filterBySize(object, min, max)
}
\arguments{
\item{object}{An object}

\item{min}{Integer, minimum size}

\item{max}{Integer, maximum size}
}
\description{
Filter by size
}
