% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersMethod.R
\name{fishersMethod}
\alias{fishersMethod}
\title{Fisher's method to combine multiple p-values}
\usage{
fishersMethod(p, returnValidp = FALSE)
}
\arguments{
\item{p}{Numeric vector, p values to be combined}

\item{returnValidP}{Logical, whether the valid p-values used should be returned as part of the list}
}
\value{
A \code{FisherMethodResult} S3 object, a list of following elements
\enumerate{
\item chisq: Chi-square statistic
\item df: Degree of freedom (which is twice the count of the valid p-values used for calculation)
\item p: p-value
\item validp (optional): valid p-values used for the calculation
}
The function returns the combined p-value using the sum of logs (Fisher's) method
}
\description{
Fisher's method to combine multiple p-values
}
\note{
The function was adapted from metap::sumlog
}
\examples{
ps <- c(0.05, 0.75)
fishersMethod(ps)
fishersMethod(ps, returnValidp=TRUE)
}
