% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\docType{methods}
\name{fisherTest,character,character,character,ANY,ANY-method}
\alias{fisherTest,character,character,character,ANY,ANY-method}
\title{Perform Fisher's exact test on a gene set}
\usage{
\S4method{fisherTest}{character,character,character,ANY,ANY}(genes, genesets,
  universe, gsName, gsCategory)
}
\arguments{
\item{genes}{a collection of genes of which over-representation of the gene set is tested}

\item{universe}{universe of genes}

\item{gsName}{gene set name, can be left missing}

\item{gsCategory}{gene set category name, can be left missing

This function performs one-sided Fisher's exact test to test the over-representation of gene set genes in the input gene list.}

\item{geneSetGenes}{genes belonging to a gene set}
}
\description{
Perform Fisher's exact test on a gene set
}
\note{
Duplicated items in genes, genesets' genes, and the universe are removed
}
\examples{
myGenes <- LETTERS[1:3]
myGeneSet1 <- LETTERS[1:6]
myGeneSet2 <- LETTERS[4:7]
myUniverse <- LETTERS
fisherTest(myGenes, myGeneSet1, myUniverse)
fisherTest(myGenes, myGeneSet2, myUniverse)
fisherTest(myGenes, myGeneSet1, myUniverse, gsName="My gene set1", gsCategory="Letters")

## note that duplicated items are removed
fisherTest(rep(myGenes,2), myGeneSet2, myUniverse)
fisherTest(rep(myGenes,2), myGeneSet2, rep(myUniverse,2))
}
