% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fisherTest,character,character,character-method}
\alias{fisherTest,character,character,character-method}
\title{Perform Fisher's exact test on a gene set}
\usage{
\S4method{fisherTest}{character,character,character}(
  genes,
  genesets,
  universe,
  gsName,
  gsNamespace,
  makeUniqueNonNA = TRUE,
  checkUniverse = TRUE,
  useEASE = FALSE
)
}
\arguments{
\item{genes}{a collection of genes of which over-representation of the gene set is tested}

\item{genesets}{A vector of character strings, genes belonging to one gene
set.}

\item{universe}{universe of genes}

\item{gsName}{gene set name, can be left missing}

\item{gsNamespace}{gene set namespace name, can be left missing}

\item{makeUniqueNonNA}{Logical, whether genes, geneSetGenes, and universe should be filtered to remove NA and made unique. The default is set to \code{TRUE}. When the uniqueness and absence of NA is ensured, this flag can be set to \code{FALSE} to accelerate the operation.}

\item{checkUniverse}{Logical, if \code{TRUE}, then genes that are in \code{genes} but are not in \code{universe} are appended to \code{universe}}

\item{useEASE}{Logical, whether to use the EASE method to report the p-value. 

This function performs one-sided Fisher's exact test to test the over-representation of gene set genes in the input gene list.

If \code{useEASE} is \code{TRUE}, one gene is penalized (removed) within \code{geneSetGenes} that are in \code{genes} and calculating the resulting Fisher exact probability for that namespace. The theoretical basis of the EASE score lies in the concept of jackknifing a probability. See Hosack \emph{et al.} for details.}
}
\description{
Perform Fisher's exact test on a gene set
}
\note{
Duplicated items in genes, genesets' genes, and the universe are per default removed
}
\examples{
myGenes <- LETTERS[1:3]
myGeneSet1 <- LETTERS[1:6]
myGeneSet2 <- LETTERS[4:7]
myUniverse <- LETTERS
fisherTest(genes=myGenes, genesets=myGeneSet1, universe=myUniverse)
fisherTest(genes=myGenes, genesets=myGeneSet2, universe=myUniverse)
fisherTest(genes=myGenes, genesets=myGeneSet1, universe=myUniverse, 
           gsName="My gene set1", gsNamespace="Letters")

## note that duplicated items are removed by default
resWoRp <- fisherTest(genes=rep(myGenes,2), genesets=myGeneSet1, 
                      universe=myUniverse)
resWithRp <- fisherTest(genes=rep(myGenes,2), genesets=myGeneSet1, 
                      universe=rep(myUniverse,2))
identical(resWoRp, resWithRp)

resWithRpNoUnique <- fisherTest(genes=rep(myGenes,2), genesets=myGeneSet1, 
           universe=rep(myUniverse,2), makeUniqueNonNA=FALSE)
identical(resWoRp, resWithRpNoUnique)
}
\references{
\describe{
  \item{Hosack \emph{et al.}}{Hosack, Douglas A., Glynn Dennis, Brad T. Sherman, H. Clifford Lane, and Richard A. Lempicki. Identifying Biological Themes within Lists of Genes with EASE. Genome Biology 4 (2003): R70. \url{https://doi.org/10.1186/gb-2003-4-10-r70}}
}
}
