% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\docType{methods}
\name{fisherTest,character,GeneSet,character,missing,missing-method}
\alias{fisherTest,character,GeneSet,character,missing,missing-method}
\title{Perform Fisher's exact test on a gene set}
\usage{
\S4method{fisherTest}{character,GeneSet,character,missing,missing}(genes,
  genesets, universe)
}
\arguments{
\item{genes}{a collection of genes of which over-representation of the gene set is tested}

\item{universe}{universe of genes
#'
This function performs one-sided Fisher's exact test to test the over-representation of gene set genes in the input gene list.}

\item{geneSetGenes}{A GeneSet object}
}
\description{
Perform Fisher's exact test on a gene set
}
\examples{
myGenes <- LETTERS[1:3]
myGeneSet1 <- new("GeneSet", category="My category 1", name="GeneSet1", genes=LETTERS[1:6])
myGeneSet2 <- new("GeneSet", category="My category 2", name="GeneSet1", genes=LETTERS[2:7])
myUniverse <- LETTERS
fisherTest(myGenes, myGeneSet1, myUniverse)
fisherTest(myGenes, myGeneSet2, myUniverse)
}
