% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\docType{methods}
\name{fisherTest,character,GeneSets,character-method}
\alias{fisherTest,character,GeneSets,character-method}
\title{Perform Fisher's exact test on a GeneSets object}
\usage{
\S4method{fisherTest}{character,GeneSets,character}(genes, genesets,
  universe, makeUniqueNonNA = TRUE, checkUniverse = TRUE,
  useEASE = FALSE)
}
\arguments{
\item{genes}{character strings of gene list to be tested}

\item{genesets}{An GeneSets object}

\item{universe}{Universe (background) gene list}

\item{makeUniqueNonNA}{Logical, whether genes and universe should be filtered to remove NA and made unique. The default is set to \code{TRUE}. When the uniqueness and absence of NA is ensured, this flag can be set to \code{FALSE} to accelerate the operation.}

\item{checkUniverse}{Logical, if \code{TRUE}, then genes that are in \code{genes} but are not in \code{universe} are appended to \code{universe}}

\item{useEASE}{Logical, whether to use the EASE method to report the p-value. 

#' @return A \code{data.table} containing Fisher's exact test results of all gene-sets, in the same order as the input gene-sets, with following columns:
\enumerate{
  \item GeneSetCategory
  \item GeneSetName
  \item GeneSetEffectiveSize, the count of genes in the gene-set that are found in the universe
  \item HitCount, the count of genes in the \code{genes} input that are in the gene-set
  \item Hits, a vector of character string, representing hits
  \item PValue
  \item FDR, PValue adjusted by the Benjamini-Hochberg method. If more than one gene-set categories are provided, the FDR correction is performed per category
}}
}
\description{
Perform Fisher's exact test on a GeneSets object
}
\examples{
gs1 <- new("GeneSet", category="A", name="GeneSet1", genes=LETTERS[1:4])
gs2 <- new("GeneSet", category="A", name="GeneSet2", genes=LETTERS[5:8])
gs3 <- new("GeneSet", category="A", name="GeneSet3", genes=LETTERS[seq(2,8,2)])
gs4 <- new("GeneSet", category="B", name="GeneSet4", genes=LETTERS[seq(1,7,2)])
gss <- GeneSets(list(gs1, gs2, gs3, gs4))
myInput <- LETTERS[2:6]
myUniverse <- LETTERS
myFisherRes <- fisherTest(myInput, gss, myUniverse)

## Note that the multiple adjustment is done by categories
gs5 <- new("GeneSet", name="GeneSet5", genes=LETTERS[1:4])
gs6 <- new("GeneSet", name="GeneSet6", genes=LETTERS[5:8])
gs7 <- new("GeneSet", name="GeneSet7", genes=LETTERS[seq(2,8,2)])
gs8 <- new("GeneSet",  name="GeneSet8", genes=LETTERS[seq(1,7,2)])
gss2 <- GeneSets(list(gs5, gs6, gs7, gs8))
myFisherRes2 <- fisherTest(myInput, gss2, myUniverse)

stopifnot(identical(myFisherRes$PValue, myFisherRes2$PValue))
stopifnot(!identical(p.adjust(myFisherRes$PValue, "fdr"), myFisherRes$FDR))
stopifnot(identical(p.adjust(myFisherRes2$PValue, "fdr"), myFisherRes2$FDR))
}
