% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fisher.R
\docType{methods}
\name{fisherTest,character,GeneSets,character,missing,missing-method}
\alias{fisherTest,character,GeneSets,character,missing,missing-method}
\title{fisherTest}
\usage{
\S4method{fisherTest}{character,GeneSets,character,missing,missing}(genes,
  genesets, universe)
}
\arguments{
\item{genes}{character strings of gene list to be tested}

\item{genesets}{An GeneSets object}

\item{universe}{Universe (background) gene list}
}
\description{
fisherTest
}
\examples{
gs1 <- new("GeneSet", category="A", name="GeneSet1", genes=LETTERS[1:4])
gs2 <- new("GeneSet", category="A", name="GeneSet2", genes=LETTERS[5:8])
gs3 <- new("GeneSet", category="A", name="GeneSet3", genes=LETTERS[seq(2,8,2)])
gs4 <- new("GeneSet", category="B", name="GeneSet4", genes=LETTERS[seq(1,7,2)])
gss <- GeneSets(list(gs1, gs2, gs3, gs4))
myInput <- LETTERS[2:6]
myUniverse <- LETTERS
myFisherRes <- fisherTest(myInput, gss, myUniverse)
## Note that the multiple adjustment is done by categories
gs5 <- new("GeneSet", name="GeneSet5", genes=LETTERS[1:4])
gs6 <- new("GeneSet", name="GeneSet6", genes=LETTERS[5:8])
gs7 <- new("GeneSet", name="GeneSet7", genes=LETTERS[seq(2,8,2)])
gs8 <- new("GeneSet",  name="GeneSet8", genes=LETTERS[seq(1,7,2)])
gss2 <- GeneSets(list(gs5, gs6, gs7, gs8))
myFisherRes2 <- fisherTest(myInput, gss2, myUniverse)#'
stopifnot(!identical(p.adjust(pValue(myFisherRes), "fdr"), fdrValue(myFisherRes)))
stopifnot(identical(p.adjust(pValue(myFisherRes2), "fdr"), fdrValue(myFisherRes2)))
}

