% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisherTest,character,GmtList,character-method}
\alias{fisherTest,character,GmtList,character-method}
\title{Perform Fisher's exact test on a GmtList object}
\usage{
\S4method{fisherTest}{character,GmtList,character}(
  genes,
  genesets,
  universe,
  gsNamespace,
  makeUniqueNonNA = TRUE,
  checkUniverse = TRUE,
  useEASE = FALSE
)
}
\arguments{
\item{genes}{character strings of gene list to be tested}

\item{genesets}{An GmtList object}

\item{universe}{Universe (background) gene list}

\item{gsNamespace}{Character string, gene-set namespace(s)}

\item{makeUniqueNonNA}{Logical, whether genes and universe should be filtered to remove NA and made unique. The default is set to \code{TRUE}. When the uniqueness and absence of NA is ensured, this flag can be set to \code{FALSE} to accelerate the operation.}

\item{checkUniverse}{Logical, if \code{TRUE}, then genes that are in \code{genes} but are not in \code{universe} are appended to \code{universe}}

\item{useEASE}{Logical, whether to use the EASE method to report the p-value.}
}
\value{
A \code{data.table} containing Fisher's exact test results of all gene-sets, in the same order as the input gene-sets, with following columns:
\enumerate{
  \item GeneSetNamespace
  \item GeneSetName
  \item GeneSetEffectiveSize, the count of genes in the gene-set that are found in the universe
  \item HitCount, the count of genes in the \code{genes} input that are in the gene-set
  \item Hits, a vector of character string, representing hits
  \item PValue
  \item FDR, PValue adjusted by the Benjamini-Hochberg method. If more than one gene-set categories are provided, the FDR correction is performed per namespace
}
}
\description{
Perform Fisher's exact test on a GmtList object
}
\examples{
gs1 <- list(name="GeneSet1", desc="desc", genes=LETTERS[1:4], namespace="A")
gs2 <- list(name="GeneSet2", desc="desc", genes=LETTERS[5:8], namespace="A")
gs3 <- list(name="GeneSet3", desc="desc", genes=LETTERS[seq(2,8,2)], namespace="A")
gs4 <- list(name="GeneSet3", desc="desc", genes=LETTERS[seq(1,7,2)], namespace="B")
gmtList <- BioQC::GmtList(list(gs1, gs2, gs3, gs4))
myInput <- LETTERS[2:6]
myUniverse <- LETTERS
myFisherRes <- fisherTest(myInput, gmtList, myUniverse)
}
