% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{fisherTest,character,list,character-method}
\alias{fisherTest,character,list,character-method}
\title{Perform Fisher's exact test on a GeneSet object}
\usage{
\S4method{fisherTest}{character,list,character}(
  genes,
  genesets,
  universe,
  makeUniqueNonNA = TRUE,
  checkUniverse = TRUE,
  useEASE = FALSE
)
}
\arguments{
\item{genes}{a collection of genes of which over-representation of the gene set is tested}

\item{genesets}{A \code{GmtList} object.}

\item{universe}{universe of genes}

\item{makeUniqueNonNA}{Logical, whether genes and universe should be filtered to remove NA and made unique. The default is set to \code{TRUE}. When the uniqueness and absence of NA is ensured, this flag can be set to \code{FALSE} to accelerate the operation.}

\item{checkUniverse}{Logical, if \code{TRUE}, then genes that are in \code{genes} but are not in \code{universe} are appended to \code{universe}}

\item{useEASE}{Logical, whether to use the EASE method to report the p-value. 

This function performs one-sided Fisher's exact test to test the over-representation of gene set genes in the input gene list.}
}
\description{
Perform Fisher's exact test on a GeneSet object
}
\examples{
myGenes <- LETTERS[1:3]
myS4GeneSet1 <- list(name="GeneSet1", desc="GeneSet", 
    genes=LETTERS[1:6], namespace="My namespace 1")
myS4GeneSet2 <- list(name="GeneSet1", desc="GeneSet", 
    genes=LETTERS[2:7], namespace="My namespace 2")
myUniverse <- LETTERS
fisherTest(myGenes, myS4GeneSet1, myUniverse)
fisherTest(myGenes, myS4GeneSet2, myUniverse)
}
