% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisherTestEdgeResult}
\alias{fisherTestEdgeResult}
\title{Run Fisher's exact test on an EdgeResult object}
\usage{
fisherTestEdgeResult(edgeResult, gmtList, contrast, thr.abs.logFC = 1,
  thr.FDR = 0.05, minGeneSetEffectiveSize = 5,
  maxGeneSetEffectiveSize = 500, ...)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{gmtList}{A \code{GmtList} or \code{GeneSets} object}

\item{contrast}{Character, the contrast of interest}

\item{thr.abs.logFC}{Numeric, threshold of absolute log2 fold-change to define positively and negatively regulated genes}

\item{thr.FDR}{Numeric, threshold of FDR values}

\item{minGeneSetEffectiveSize}{Integer, minimal number of genes of a geneset that are quantified}

\item{maxGeneSetEffectiveSize}{Integer, maximal number of genes of a geneset that are quantified}
}
\value{
## TODO: example
}
\description{
Run Fisher's exact test on an EdgeResult object
}
