% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{fisherTestResultNewHitsProp}
\alias{fisherTestResultNewHitsProp}
\title{Append NewHitsProp to the result \code{data.table} returned by \code{fisherTest}}
\usage{
fisherTestResultNewHitsProp(fisherTestResults)
}
\arguments{
\item{fisherTestResults}{\code{data.table} returned by \code{fisherTest}}
}
\value{
A new \code{data.table} containing all columns of the input and \code{NewHitsProp}, a new column including the proportion of new hits in the gene-set
}
\description{
Append NewHitsProp to the result \code{data.table} returned by \code{fisherTest}
}
