% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{fullEnrichTable}
\alias{fullEnrichTable}
\title{Extract the full enrichment table}
\usage{
fullEnrichTable(edgeGse)
}
\arguments{
\item{edgeGse}{An \code{EdgeGSE} object}
}
\value{
A \code{data.frame}
}
\description{
Extract the full enrichment table
}
