% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\docType{data}
\name{GeMS_GENESETS_URL}
\alias{GeMS_GENESETS_URL}
\title{GeMS genesets retrieval URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
GeMS_GENESETS_URL
}
\description{
GeMS genesets retrieval URL
}
\keyword{datasets}
