% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\docType{data}
\name{GeMS_INSERT_URL}
\alias{GeMS_INSERT_URL}
\title{GeMS insert URL}
\format{
An object of class \code{character} of length 1.
}
\usage{
GeMS_INSERT_URL
}
\description{
GeMS insert URL
}
\keyword{datasets}
