% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\docType{data}
\name{GeMS_TEST_GENESETS_URL}
\alias{GeMS_TEST_GENESETS_URL}
\title{GeMS geneset retrieval URL for testing}
\format{
An object of class \code{character} of length 1.
}
\usage{
GeMS_TEST_GENESETS_URL
}
\description{
GeMS geneset retrieval URL for testing
}
\keyword{datasets}
