% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\docType{class}
\name{GeneSetResult-class}
\alias{GeneSetResult-class}
\title{A generic, virtual S4 class for gene-set analysis result}
\description{
A generic, virtual S4 class for gene-set analysis result
}
