% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTopTable.R
\name{getFDRCol}
\alias{getFDRCol}
\title{Get the name of the column which store false-discovery rates (adjusted P-values) from topTables}
\usage{
getFDRCol(colnames)
}
\arguments{
\item{colnames}{A character string vector of column names}
}
\value{
The column name of the FDRs, \code{NA} if not found.
}
\description{
Get the name of the column which store false-discovery rates (adjusted P-values) from topTables
}
\examples{
getFDRCol(c("Feature", "logFC", "PValue", "FDR"))
getFDRCol(c("Feature", "logFC", "P.Value", "FDR"))
getFDRCol(c("Feature", "logFC", "p.Value", "adjPvalue"))
getFDRCol(c("Feature", "logFC", "PValue", "adj.PValue"))
}
