% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{getJsonResponse}
\alias{getJsonResponse}
\title{Send a list as JSON query to an URL and fetch the response}
\usage{
getJsonResponse(url, body)
}
\arguments{
\item{url}{The destination URL}

\item{body}{A list to be sent to the URL, which will be encoded in the JSON format internally}
}
\value{
The response from the webserver
}
\description{
Send a list as JSON query to an URL and fetch the response
}
\examples{
\dontrun{
   getJsonResponse(GeMS_GENESETS_URL, list(user=ribiosUtils::whoami()))
}
}
