% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseTopTable.R
\name{getPvalCol}
\alias{getPvalCol}
\title{Get the name of the column which store unadjusted P-values from topTables}
\usage{
getPvalCol(colnames)
}
\arguments{
\item{colnames}{A character string vector of column names}
}
\value{
The column name of the unadjusted p-values, \code{NA} if not found.
}
\description{
Get the name of the column which store unadjusted P-values from topTables
}
\examples{
getPvalCol(c("Feature", "logFC", "PValue", "FDR"))
getPvalCol(c("Feature", "logFC", "P.Value", "FDR"))
getPvalCol(c("Feature", "logFC", "p.Value", "adjPvalue"))
getPvalCol(c("Feature", "logFC", "pval", "adjPvalue"))
}
