% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{getSetsWithNamesFromGeMS}
\alias{getSetsWithNamesFromGeMS}
\title{Get one or more gene-sets with their names}
\usage{
getSetsWithNamesFromGeMS(setNames = NULL)
}
\arguments{
\item{setNames}{Character strings}
}
\value{
A GmtList object
}
\description{
Get one or more gene-sets with their names
}
\examples{
\dontrun{
getSetsWithNamesFromGeMS(c("Plasma_sc", "Bcell_l_Danaher17"))
}
}
\seealso{
\code{\link{getSetWithNameFromGeMS}}
}
