% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{getSetsWithPropertyFromGeMS}
\alias{getSetsWithPropertyFromGeMS}
\title{Get gene-sets for application}
\usage{
getSetsWithPropertyFromGeMS(property = "meta.application", value = "")
}
\arguments{
\item{property}{Character string, property to query}

\item{value}{Character string, property value}
}
\value{
A GmtList object
}
\description{
Get gene-sets for application
}
\examples{
\dontrun{
getSetsWithPropertyFromGeMS("meta.application", "rtbeda_CIT")
}
}
