% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{getSetWithNameFromGeMS}
\alias{getSetWithNameFromGeMS}
\title{Get one gene-set with its name}
\usage{
getSetWithNameFromGeMS(setName)
}
\arguments{
\item{setName}{Character string}
}
\value{
A list of two elements
\enumerate{
  \item name
  \item genes
}
}
\description{
Get one gene-set with its name
}
\examples{
\dontrun{
getSetWithNameFromGeMS("Plasma_sc")
}
}
\seealso{
\code{\link{getSetsWithNamesFromGeMS}}
}
