% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{getUserSetsFromGeMS}
\alias{getUserSetsFromGeMS}
\title{Get gene sets of a user from GeMS}
\usage{
getUserSetsFromGeMS(user = ribiosUtils::whoami())
}
\arguments{
\item{user}{User name}
}
\value{
A data.frame including following columns: 
\enumerate{
  \item setName
  \item desc
  \item domain
  \item source
  \item subtype
}
}
\description{
Get gene sets of a user from GeMS
}
\examples{
\dontrun{
## my gene-sets
getUserSetsFromGeMS()
## from another user
getUserSetsFromGeMS("kanga6")
}
}
