% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/camera.R
\name{gmtListCamera}
\alias{gmtListCamera}
\title{CAMERA applied to GmtList}
\usage{
gmtListCamera(matrix, geneSymbols, gmtList, design, contrasts)
}
\arguments{
\item{matrix}{Expression matrix, features (genes) in rows and samples in columns}

\item{geneSymbols}{Gene-symbols corresponding to the rows of matrix}

\item{gmtList}{A \code{\link[BioQC]{GmtList}} object}

\item{design}{Design matrix}

\item{contrasts}{Contrast matrix}
}
\value{
A \code{data.frame} containing CAMERA results
}
\description{
CAMERA applied to GmtList
}
\examples{
mat <- matrix(c(10,3,5,9,3,5,
                2,4,8,12,9,9,
                3,5,7,5,4,4,
                3,3,12,12,0,1), ncol=6, byrow=TRUE)
rownames(mat) <- sprintf("gene\%d", 1:nrow(mat))
designMatrix <- matrix(c(rep(1,6), c(0,0,1,1,0,0), c(0,0,0,0,1,1)),
                       byrow=FALSE, ncol=3)
contrastMatrix <- matrix(c(0,1,0,0,0,1), ncol=2, byrow=FALSE)
gs1 <- list(name="GeneSet1", desc="", genes=c("gene1", "gene3"), namespace="default")
gs2 <- list(name="GeneSet2", desc="", genes=c("gene2", "gene4"), namespace="default")
gs3 <- list(name="GeneSet3", desc="", genes=c("gene1", "gene4"), namespace="default")
gmtlist <- BioQC::GmtList(list(gs1, gs2, gs3))
cameraOut <- gmtListCamera(mat, rownames(mat), gmtlist, designMatrix, contrastMatrix)
cameraOut
}
