% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gsDesc}
\alias{gsDesc}
\title{Return gene-set description}
\usage{
gsDesc(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set description
}
