% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosGSEA-package.R
\docType{data}
\name{GSEA_DATA_DIR}
\alias{GSEA_DATA_DIR}
\alias{GSEA_ANNOTATION_DIR}
\alias{GSEA_GENESET_DIR}
\alias{JAVA_BIN}
\alias{GSEA_JAR}
\alias{DEFAULT_GMT}
\alias{DEFAULT_CHIP}
\title{Variables useful for GSEA analysis in Roche bioinformatics environment.}
\description{
These variables ar useful for performing GSEA analysis in Roche
bioinformatics environment.
}
