% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaCoreEnrichGenes}
\alias{gseaCoreEnrichGenes}
\alias{gseaLeadingEdgeGenes}
\alias{gseaCoreEnrichGenes,annoBroadGseaResItem-method}
\alias{gseaCoreEnrichGenes,annoBroadGseaRes-method}
\title{Return GSEA core enrichment genes (also known as leading-edge genes)}
\usage{
gseaCoreEnrichGenes(object)

\S4method{gseaCoreEnrichGenes}{annoBroadGseaResItem}(object)

\S4method{gseaCoreEnrichGenes}{annoBroadGseaRes}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA core enrichment genes (also known as leading-edge genes)
}
\section{Methods (by class)}{
\itemize{
\item \code{annoBroadGseaResItem}: Return core enriched genes (also known as
leading-edge genes) in an annoBroadGseaResItem object as a character string
vector.

\item \code{annoBroadGseaRes}: Return core enriched genes (also known as
leading-edge genes) in an annoBroadGseaRes object as a list of character 
string vectors.
}}

