% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gseaCoreEnrichThr}
\alias{gseaCoreEnrichThr}
\title{Return GSEA core enrichment score threshold}
\usage{
gseaCoreEnrichThr(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA core enrichment score threshold
}
