% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaFDR}
\alias{gseaFDR}
\alias{gseaFDR,broadGseaResItem-method}
\alias{gseaFDR,annoBroadGseaRes-method}
\alias{gseaFDR,annoBroadGseaResList-method}
\title{Return GSEA FDR}
\usage{
gseaFDR(object)

\S4method{gseaFDR}{broadGseaResItem}(object)

\S4method{gseaFDR}{annoBroadGseaRes}(object)

\S4method{gseaFDR}{annoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA FDR
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get GSEA FDR values
from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get GSEA FDR values
from an annoBroadGseaRes object

\item \code{annoBroadGseaResList}: Get GSEA FDR values
from an annoBroadGseaResList object
}}

