% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{gseaFingerprintMatrix}
\alias{gseaFingerprintMatrix}
\title{Return a fingerprint matrix of a vector of GSEA output directories}
\usage{
gseaFingerprintMatrix(gseaDirs, value = c("q", "es", "nes"), ...)
}
\arguments{
\item{gseaDirs}{A vector of character strings, paths to GSEA output
directories}

\item{value}{Type of value to build the matrix, passed to
\code{gseaFingerPrint}}
}
\value{
A matrix
}
\description{
Return a fingerprint matrix of a vector of GSEA output directories
}
