% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaNES}
\alias{gseaNES}
\alias{gseaNES,broadGseaResItem-method}
\alias{gseaNES,annoBroadGseaRes-method}
\alias{gseaNES,annoBroadGseaResList-method}
\title{Return GSEA normalized enrichment scores}
\usage{
gseaNES(object)

\S4method{gseaNES}{broadGseaResItem}(object)

\S4method{gseaNES}{annoBroadGseaRes}(object)

\S4method{gseaNES}{annoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA normalized enrichment scores
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get GSEA normalized enrichment score 
from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get GSEA normalized enrichment score 
from an annoBroadGseaRes object

\item \code{annoBroadGseaResList}: Get GSEA normalized enrichment score 
from an annoBroadGseaResList object
}}

