% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaNP}
\alias{gseaNP}
\alias{gseaNP,broadGseaResItem-method}
\alias{gseaNP,annoBroadGseaRes-method}
\alias{gseaNP,annoBroadGseaResList-method}
\title{Return GSEA number of permutation}
\usage{
gseaNP(object)

\S4method{gseaNP}{broadGseaResItem}(object)

\S4method{gseaNP}{annoBroadGseaRes}(object)

\S4method{gseaNP}{annoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA number of permutation
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get GSEA number of permutations
from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get GSEA number of permutations
from an annoBroadGseaRes object

\item \code{annoBroadGseaResList}: Get GSEA number of permutations
from an annoBroadGseaResList object
}}

