% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fingerprint.R
\name{gseaResQvalue}
\alias{gseaResQvalue}
\title{Extract Q-values from GSEA result file}
\usage{
gseaResQvalue(file, threshold = 1e-04, log = FALSE, posLog = FALSE)
}
\arguments{
\item{file}{GSEA result file}

\item{threshold}{Numeric, threshold}

\item{log}{Logical, whether the Q-value should be reported as
log10-transformed value.}

\item{posLog}{Logical, whether the Q-value should be reported as absolute of
the log10-transformed value.}
}
\description{
Extract Q-values from GSEA result file
}
