% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsEffectiveSize}
\alias{gsEffectiveSize}
\alias{gsEffectiveSize,FisherResult-method}
\alias{gsEffectiveSize,FisherResultList-method}
\title{Return the effective size of gene-set}
\usage{
gsEffectiveSize(object, ...)

\S4method{gsEffectiveSize}{FisherResult}(object)

\S4method{gsEffectiveSize}{FisherResultList}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return the effective size of gene-set
}
\section{Methods (by class)}{
\itemize{
\item \code{gsEffectiveSize(FisherResult)}: Effective sizes of gene-set, returning an integer.

\item \code{gsEffectiveSize(FisherResultList)}: Effective sizes of Gene-sets, returning an integer vector.

}}
