% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsEffSize}
\alias{gsEffSize}
\alias{gsEffSize,FisherResult-method}
\alias{gsEffSize,FisherResultList-method}
\title{Return gene-set effect size}
\usage{
gsEffSize(object, ...)

\S4method{gsEffSize}{FisherResult}(object)

\S4method{gsEffSize}{FisherResultList}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{



}
\description{
Return gene-set effect size
}
\section{Methods (by class)}{
\itemize{
\item \code{FisherResult}: Gene-set effect size, returning a numeric value.

\item \code{FisherResultList}: Gene-set effect size, returning a numeric vector.
}}

