% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{gsFisherTest}
\alias{gsFisherTest}
\title{Perform Fisher's exact test on a gene set}
\usage{
gsFisherTest(genes, geneSetGenes, universe, gsName, gsCategory)
}
\arguments{
\item{genes}{a collection of genes of which over-representation of the gene set is tested}

\item{geneSetGenes}{genes belonging to a gene set}

\item{universe}{universe of genes}

\item{gsName}{gene set name}

\item{gsCategory}{gene set category name

This function performs one-sided Fisher's exact test to test the over-representation of gene set genes in the input gene list. Note that all three parameters must not contain duplicates! The S4-class methods ensure this by checking the input parameters.}
}
\description{
Perform Fisher's exact test on a gene set
}
\examples{
myGenes <- LETTERS[1:3]
myGeneSet1 <- LETTERS[1:6]
myGeneSet2 <- LETTERS[4:7]
myUniverse <- LETTERS
ribiosGSEA:::gsFisherTest(myGenes, myGeneSet1, myUniverse)
ribiosGSEA:::gsFisherTest(myGenes, myGeneSet2, myUniverse)
ribiosGSEA:::gsFisherTest(myGenes, myGeneSet1, myUniverse, gsName="My gene set1", gsCategory="Letters")
}
