% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gsGeneCount}
\alias{gsGeneCount}
\title{Return gene-set gene count}
\usage{
gsGeneCount(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set gene count
}
