% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGeneIndices}
\alias{gsGeneIndices}
\alias{gsGeneIndices,BroadGseaResItem-method}
\title{Return gene-set gene indices}
\usage{
gsGeneIndices(object)

\S4method{gsGeneIndices}{BroadGseaResItem}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return gene-set gene indices
}
\section{Methods (by class)}{
\itemize{
\item \code{BroadGseaResItem}: Get gene-set gene indices
from a BroadGseaResItem object, returning a vector of integers.
}}

