% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGenes}
\alias{gsGenes}
\alias{gsGenes,annoBroadGseaResItem-method}
\alias{gsGenes,annoBroadGseaRes-method}
\alias{gsGenes,GmtList-method}
\title{Return gene-set genes}
\usage{
gsGenes(object, ...)

\S4method{gsGenes}{annoBroadGseaResItem}(object)

\S4method{gsGenes}{annoBroadGseaRes}(object)

\S4method{gsGenes}{GmtList}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set genes
}
\section{Methods (by class)}{
\itemize{
\item \code{annoBroadGseaResItem}: Get gene-set genes
from a broadGseaResItem object, returning a character string vector.

\item \code{annoBroadGseaRes}: Get gene-set genes from an annoBroadGseaRes object,
returning a list of character string vectors.

\item \code{GmtList}: Get gene-set genes from a GmtList object, returning a
list of character string vector. It uses the implementation in BioQC.
}}

