% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gsGeneValues<-}
\alias{gsGeneValues<-}
\title{Set gene-set gene statistics (values)}
\usage{
gsGeneValues(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Value}
}
\description{
Set gene-set gene statistics (values)
}
