% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGeneValues}
\alias{gsGeneValues}
\alias{gsGeneValues,AnnoBroadGseaResItem-method}
\alias{gsGeneValues,AnnoBroadGseaRes-method}
\title{Return gene-set gene values}
\usage{
gsGeneValues(object)

\S4method{gsGeneValues}{AnnoBroadGseaResItem}(object)

\S4method{gsGeneValues}{AnnoBroadGseaRes}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return gene-set gene values
}
\section{Methods (by class)}{
\itemize{
\item \code{AnnoBroadGseaResItem}: Return values associated with the genes in a
gene-set in an AnnoBroadGseaResItem object in a numeric vector.

\item \code{AnnoBroadGseaRes}: Return values associated with the genes in a
gene-set in an AnnoBroadGseaRes object in a list of numeric vectors.
}}

