% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsName}
\alias{gsName}
\alias{gsName,BroadGseaResItem-method}
\alias{gsName,AnnoBroadGseaRes-method}
\alias{gsName,FisherResult-method}
\alias{gsName,FisherResultList-method}
\alias{gsName,GmtList-method}
\title{Return gene-set name}
\usage{
gsName(object, ...)

\S4method{gsName}{BroadGseaResItem}(object)

\S4method{gsName}{AnnoBroadGseaRes}(object)

\S4method{gsName}{FisherResult}(object)

\S4method{gsName}{FisherResultList}(object, ...)

\S4method{gsName}{GmtList}(object)

\S4method{gsName}{FisherResultList}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set name
}
\section{Methods (by class)}{
\itemize{
\item \code{gsName(BroadGseaResItem)}: Get gene-set name from a BroadGseaResItem object

\item \code{gsName(AnnoBroadGseaRes)}: Get gene-set name from an AnnoBroadGseaRes object

\item \code{gsName(FisherResult)}: Get gene-set name from a FisherResult object

\item \code{gsName(FisherResultList)}: Get gene-set name from a FisherResultList object

\item \code{gsName(GmtList)}: Get gene-set name from a GmtList object

\item \code{gsName(FisherResultList)}: Get gene-set name from a FisherResultList object

}}
