% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsNamespace}
\alias{gsNamespace}
\alias{gsNamespace,GmtList-method}
\alias{gsNamespace,FisherResult-method}
\alias{gsNamespace,FisherResultList-method}
\title{Return gene-set namespace}
\usage{
gsNamespace(object, ...)

\S4method{gsNamespace}{GmtList}(object)

\S4method{gsNamespace}{FisherResult}(object)

\S4method{gsNamespace}{FisherResultList}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set namespace
}
\section{Methods (by class)}{
\itemize{
\item \code{GmtList}: Return gene-set namespace from a GmtList object

\item \code{FisherResult}: Return gene-set namespace from a FisherResult object

\item \code{FisherResultList}: Return gene-set namespace from a FisherResultList 
object.
}}

