% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gsSize}
\alias{gsSize}
\title{Return the size (unique length) of gene-sets}
\usage{
gsSize(gmtList)
}
\arguments{
\item{gmtList}{a \code{GmtList} object}
}
\value{
An integer vector
}
\description{
Return the size (unique length) of gene-sets
}
