% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{hits}
\alias{hits}
\alias{hits,FisherResult-method}
\alias{hits,FisherResultList-method}
\title{Return hits}
\usage{
hits(object, ...)

\S4method{hits}{FisherResult}(object)

\S4method{hits}{FisherResultList}(object, geneset)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}

\item{geneset}{Character string, gene-set name}
}
\description{
Return hits
}
\section{Methods (by class)}{
\itemize{
\item \code{FisherResult}: Return hits from a FisherResult object

\item \code{FisherResultList}: Return hits from a FisherResultList object, returning a list
if \code{geneset} is missing, or gene-set genes if \code{geneset} is present.
}}

