% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{insertGmtListToGeMS}
\alias{insertGmtListToGeMS}
\title{Insert a GmtList object to GeMS}
\usage{
insertGmtListToGeMS(gmtList, geneFormat = 0, source = "PubMed",
  xref = "", taxID = 9606, user = ribiosUtils::whoami())
}
\arguments{
\item{gmtList}{A \code{GmtList} object defined in the \code{BioQC} package}

\item{geneFormat}{Integer index of gene format. 0 stands for official human gene symbol}

\item{source}{Character, source of the gene set}

\item{xref}{External reference, will be written in the meta field of the backend database}

\item{taxID}{Integer, NCBI taxonomy ID of the species.}

\item{user}{The user name}
}
\value{
Response code or error message returned by the GeMS API. A value of \code{200} indicates a successful insertion.
}
\description{
Insert a GmtList object to GeMS
}
\examples{
\dontrun{
  testList <- list(list(name="GS_A", desc=NULL, genes=c("MAPK14", "JAK1", "EGFR")),
    list(name="GS_B", desc="gene set B", genes=c("ABCA1", "DDR1", "DDR2")),
    list(name="GS_C", desc="gene set C", genes=NULL))
  testGmt <- BioQC::GmtList(testList)
  insertGmtListToGeMS(testGmt, geneFormat=0, source="Test", xref="PMID:000000")
  removeFromGeMS(setName=c("GS_A", "GS_B", "GS_C"), source="Test")
}
}
\seealso{
\code{\link{removeFromGeMS}}
}
