% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{insertGmtListToGeMSBody}
\alias{insertGmtListToGeMSBody}
\title{Construct message body to insert into GeMS}
\usage{
insertGmtListToGeMSBody(gmtList, geneFormat = 0, source = "PubMed",
  xref = "", taxID = 9606, user = ribiosUtils::whoami(),
  subtype = "", domain = "")
}
\arguments{
\item{gmtList}{A \code{GmtList} object defined in the \code{BioQC} package}

\item{geneFormat}{Integer index of gene format. 0 stands for official human gene symbol}

\item{source}{Character, source of the gene set}

\item{xref}{External reference, will be written in the meta field of the backend database}

\item{taxID}{Integer, NCBI taxonomy ID of the species.}

\item{user}{The user name}

\item{subtype}{Subtype of the geneset}

\item{domain}{Domain of the geneset}
}
\value{
A list with three items: \code{headers}, \code{parsed}, and \code{params}
}
\description{
Construct message body to insert into GeMS
}
\examples{
testList <- list(list(name="GS_A", desc=NULL, genes=c("MAPK14", "JAK1", "EGFR")),
  list(name="GS_B", desc="gene set B", genes=c("ABCA1", "DDR1", "DDR2")),
  list(name="GS_C", desc="gene set C", genes=NULL))
testGmt <- BioQC::GmtList(testList)
insertGmtListToGeMSBody(testGmt, geneFormat=0, source="Test", xref="PMID:000000")
}
