% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{isGeMSReachable}
\alias{isGeMSReachable}
\title{Test whether GeMS is reachable}
\usage{
isGeMSReachable()
}
\value{
Logical value
}
\description{
Test whether GeMS is reachable
}
\examples{
\dontrun{
isGeMSReachble()
}
}
