% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{isGseaCoreEnrich}
\alias{isGseaCoreEnrich}
\title{Returns a logical value whether the enrichment belongs to the core}
\usage{
isGseaCoreEnrich(object)
}
\arguments{
\item{object}{An object}
}
\description{
Returns a logical value whether the enrichment belongs to the core
}
