% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{isSigGeneSet}
\alias{isSigGeneSet}
\title{Return a logical vector indicating whether a gene-set is significantly
enriched}
\usage{
isSigGeneSet(object, fdr, ...)
}
\arguments{
\item{object}{An object}

\item{fdr}{Numeric, FDR threshold}

\item{...}{Other parameters}
}
\description{
Return a logical vector indicating whether a gene-set is significantly
enriched
}
