% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallW.R
\name{kendallW.eSet}
\alias{kendallW.eSet}
\title{Compute Kendall's W for an eSet object}
\usage{
\method{kendallW}{eSet}(
  object,
  row.factor,
  summary = c("none", "mean", "median", "max.mean.sig", "max.var.sig"),
  na.rm = TRUE,
  alpha = 0.01
)
}
\arguments{
\item{object}{row.factor Factor vector of the same length as the row counts}

\item{summary}{Summary type, passed to \code{kendallWmat}}

\item{na.rm}{Logical, whether \code{NA} values should be removed}

\item{alpha}{Numeric, passed to \code{kendallWmat}}
}
\description{
Compute Kendall's W for an eSet object
}
\seealso{
\code{\link{kendallWmat}}
}
