% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallW.R
\name{kendallW}
\alias{kendallW}
\title{S3 generic for kendallW}
\usage{
kendallW(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{
A matrix with an \code{info} attribute, see \code{\link{kendallWmat}}.
}
\description{
S3 generic for kendallW
}
