% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallW.R
\name{kendallWinfo<-.matrix}
\alias{kendallWinfo<-.matrix}
\title{S3 method to assign kendallW information to a matrix}
\usage{
\method{kendallWinfo}{matrix}(object) <- value
}
\arguments{
\item{object}{matrix}

\item{value}{assigned value}
}
\value{
The \code{matrix} containing grouping information
}
\description{
S3 method to assign kendallW information to a matrix
}
