% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendallW.R
\name{kendallWinfo}
\alias{kendallWinfo}
\alias{kendallWinfo.matrix}
\title{S3 generic for kendallW information}
\usage{
kendallWinfo(object)

\method{kendallWinfo}{matrix}(object)
}
\arguments{
\item{object}{An object}
}
\value{
A \code{data.frame} containing grouping information.
}
\description{
S3 generic for kendallW information
}
\section{Methods (by class)}{
\itemize{
\item \code{kendallWinfo(matrix)}: Extract kendallW information from a matrix

}}
