% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{logFCgage}
\alias{logFCgage}
\title{Perform the GAGE analysis for EdgeResult and GmtList}
\usage{
logFCgage(edgeResult, gmtList)
}
\arguments{
\item{edgeResult}{An \code{EdgeResult} object}

\item{gmtList}{A \code{GmtList} object}
}
\value{
A \code{EdgeGSE} object
}
\description{
Perform the GAGE analysis for EdgeResult and GmtList
}
