% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{mergeCameraResults}
\alias{mergeCameraResults}
\title{Merge CAMERA results using limma default parameters and biosCamera parameters}
\usage{
mergeCameraResults(
  matrix,
  index,
  designMatrix,
  contrast,
  featureLabels,
  weights = NULL,
  use.ranks = FALSE
)
}
\arguments{
\item{matrix}{A numeric matrix, passed to \code{camera} and \code{biosCamera}}

\item{index}{An index vector or a list of index vectors of features.}

\item{designMatrix}{Design matrix.}

\item{contrast}{A numeric vector of the same length as the number of columns in the design matrix, coefficients of contrasts.}

\item{featureLabels}{A character vector of the same length as the number of rows of the matrix, feature labels, for instance gene symbols.}

\item{weights}{NULL or numeric matrix of precision weights, passed to \code{camera}.}

\item{use.ranks}{Logical, passed to \code{camera}.

The function merges the output of \code{camera} with default options with
the output of \code{biosCamera}, which appends additional information to 
camera methods, to return a comprehensive table as output.}
}
\description{
Merge CAMERA results using limma default parameters and biosCamera parameters
}
\examples{

y <- matrix(rnorm(1000*6),1000,6)
features <- sprintf("Feature\%d", 1:nrow(y))
design <- cbind(Intercept=1,Group=c(0,0,0,1,1,1))
# First set of 20 genes are genuinely deferentially expressed 
index1 <- 1:20
y[index1,4:6] <- y[index1,4:6]+1
# The second set of 20 genes are not
index2 <- 21:40
index1Res <- mergeCameraResults(y, index=index1, 
  designMatrix=design, contrast=c(0,1), featureLabels=features)
index1ListRes <- mergeCameraResults(y, index=list(index1), 
   designMatrix=design, contrast=c(0,1), featureLabels=features)
index12ListRes <- mergeCameraResults(y, index=list(index1, index2), 
   designMatrix=design, contrast=c(0,1), featureLabels=features)
}
\seealso{
\code{\link{camera}}, \code{\link{biosCamera}}
}
