% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{minFDRvalue}
\alias{minFDRvalue}
\title{Return the minimal FDR value from a FisherResultList}
\usage{
minFDRvalue(object)
}
\arguments{
\item{object}{A FisherResultList object}
}
\value{
A numeric value
}
\description{
Return the minimal FDR value from a FisherResultList
}
