% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{minPValue}
\alias{minPValue}
\title{Return the minimum P value}
\usage{
minPValue(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return the minimum P value
}
