% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strWithNumber.R
\name{orderByNumberInStr}
\alias{orderByNumberInStr}
\title{Order strings by numbers in them}
\usage{
orderByNumberInStr(str, ...)
}
\arguments{
\item{str}{A vector of character trings}

\item{...}{Passed to \code{order}, by default \code{decreasing} is
\code{TRUE}, i.e. the descending order is reported.}
}
\description{
Order strings by numbers in them
}
\examples{
orderByNumberInStr(c("D1", "D10", "D15", "D3.5"))
}
