% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBroadGSEA.R
\name{parseGSEAdir}
\alias{parseGSEAdir}
\title{Parse an output directory of the Broad GSEA tool}
\usage{
parseGSEAdir(dir)
}
\arguments{
\item{dir}{Character string, path to output directory}
}
\value{
An \code{annoBroadGseaRes} object
}
\description{
Parse an output directory of the Broad GSEA tool
}
