% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyRonetGenesetNames.R
\name{prettyRonetGenesetNames}
\alias{prettyRonetGenesetNames}
\title{Pretty RONET Gene-set Names}
\usage{
prettyRonetGenesetNames(x, nchar = 50)
}
\arguments{
\item{x}{Character strings, RONET gene-set names}

\item{nchar}{Integer, number of chararacters to be displayed.}
}
\value{
Character strings
}
\description{
Pretty RONET Gene-set Names
}
\examples{
strs <- c("ARNT_GeneID405_negativeTargets",
  "Neurophysiological_process_nNOS_signaling_in_neuronal_synapses",
  "NR5A1_GeneID2516_allTargets", 
  "IL4_GeneID3565_negativeTargets",
  "Apoptosis_REACTOME")
prettyRonetGenesetNames(strs)
  
}
