% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClasses.R
\name{print.FisherResult}
\alias{print.FisherResult}
\title{Print a FisherResult object}
\usage{
\method{print}{FisherResult}(x, ...)
}
\arguments{
\item{x}{A FisherResult object}

\item{...}{Not used}
}
\description{
Print a FisherResult object
}
