% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishersMethod.R
\name{print.FishersMethodResult}
\alias{print.FishersMethodResult}
\title{Print S3 object FishersMethodResult}
\usage{
\method{print}{FishersMethodResult}(x, ...)
}
\arguments{
\item{x}{An object of the \code{FisherMethodResult} (S3) class}

\item{...}{Not used}
}
\description{
Print S3 object FishersMethodResult
}
