% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{printContributingGenes}
\alias{printContributingGenes}
\title{Print contributing genes}
\usage{
printContributingGenes(geneLabels, geneValues)
}
\arguments{
\item{geneLabels}{A vector of character strings}

\item{geneValues}{A vector of numeric values}
}
\value{
A vector of character strings
}
\description{
Print contributing genes
}
