% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{pValue}
\alias{pValue}
\alias{pValue,FisherResult-method}
\alias{pValue,FisherResultList-method}
\alias{fdrValue,FisherResult-method}
\alias{fdrValue,FisherResultList-method}
\title{Return P-values}
\usage{
pValue(object, ...)

\S4method{pValue}{FisherResult}(object)

\S4method{pValue}{FisherResultList}(object, ind, ...)

\S4method{fdrValue}{FisherResult}(object)

\S4method{fdrValue}{FisherResultList}(object, ind, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}

\item{ind}{An integer or logical vector for subsetting}
}
\description{
Return P-values
}
\section{Methods (by class)}{
\itemize{
\item \code{FisherResult}: Return the p-value from a FisherResult

\item \code{FisherResultList}: Return the p-values from a FisherResultList. If \code{ind}
is missing, all p-values are returned; otherwise, the subset indicated by
\code{ind} is returned.

\item \code{FisherResult}: Return the FDR-value from a FisherResult

\item \code{FisherResultList}: Return the FDR-values from a FisherResultList. If \code{ind}
is missing, all FDR-values are returned; otherwise, the subset indicated by
\code{ind} is returned.
}}

