% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{readDefaultGenesets}
\alias{readDefaultGenesets}
\title{Read default genesets for gene-set enrichment analysis}
\usage{
readDefaultGenesets(path, mps = FALSE)
}
\arguments{
\item{path}{Character, path to the directory where the gmt files are stored}

\item{mps}{Logical,  whether molecular-phenotypic screening (MPS) genesets should be read in as pathway-centric categories (\code{TRUE}) or as one category named \code{MolecularPhenotyping} (\code{FALSE}).}
}
\description{
In Roche Bioinformatics we use a default collection of gene-sets for gene-set enrichment analysis. This function loads this collection.
}
\details{
The default collection includes both publicly available genesets as well as proprietary genesets, and therefore they are not included as part of the ribios package.

Publicly available genesets include
\itemize{
 \item{MSigDB: collections C2, C7 and Hallmark}
 \item{RONET: which is a collection of publicly available pathway databases including REACTOME and NCI-Nature}
 \item{goslim}
}
}
\examples{
\dontrun{
  readDefaultGenesets("/tmp/defaultGmts")
}
}
