% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{readGmt}
\alias{readGmt}
\title{Read GMT file into a GeneSets object}
\usage{
readGmt(..., category = NULL)
}
\arguments{
\item{...}{Named or unnamed characater string vector, giving file names of One or more GMT format files.}

\item{category}{Character string, category of the gene sets, for instance 'Gene Onotology' or 'Reactome pathway'}
}
\description{
Read GMT file into a GeneSets object
}
\examples{
gmtFile <- system.file("extdata", "example.gmt", package="ribiosGSEA")
mySet <- readGmt(gmtFile)
myFakeSet <- readGmt(CategoryA=gmtFile, CategoryB=gmtFile)
anotherFakeSet <- readGmt(gmtFile, gmtFile, category=c("CategoryA", "CategoryB"))

}
