% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{readMPSGmt}
\alias{readMPSGmt}
\title{Read molecular-phenotyping genesets}
\usage{
readMPSGmt(file)
}
\arguments{
\item{file}{GMT file which stores default molecular-phenotyping genesets}
}
\value{
A \code{GeneSet} object containing molecular-phenotypic screening (MPS) categories and genes
}
\description{
Read molecular-phenotyping genesets
}
