% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post-processing.R
\name{readSigCameraScoreMatrix}
\alias{readSigCameraScoreMatrix}
\title{Read significant CAMERA results into a matrix}
\usage{
readSigCameraScoreMatrix(file, ...)
}
\arguments{
\item{file}{A tsv file, output of \code{\link{biosCamera}}}

\item{...}{passed to \code{readSigCameraResults}}
}
\value{
A numeric matrix with gene-sets as rows and contrasts as columns.
}
\description{
Read significant CAMERA results into a matrix
}
