% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosGSEA-package.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{DesignContrast}
\alias{EdgeObject}
\alias{DGEList}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{edgeR}{\code{\link[edgeR]{DGEList}}}

  \item{ribiosExpression}{\code{\link[ribiosExpression]{DesignContrast}}}

  \item{ribiosNGS}{\code{\link[ribiosNGS]{EdgeObject}}}
}}

