% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{removeFromGeMS}
\alias{removeFromGeMS}
\title{Rmove one or gene sets of the same source and user from GeMS}
\usage{
removeFromGeMS(setName = "", source = "",
  user = ribiosUtils::whoami(), subtype = "")
}
\arguments{
\item{setName}{A vector of character strings, defining set names to be renamed. They must all have the same \code{source}, \code{user}, and \code{subtype}}

\item{source}{Character string, source of the gene set(s)}

\item{user}{Character string, user name}

\item{subtype}{Character string, subtype of the gene set(s)}
}
\value{
Response code or error message returned by the GeMS API. A value of \code{200} indicates a successful insertion.
}
\description{
Rmove one or gene sets of the same source and user from GeMS
}
\examples{
\dontrun{
  testList <- list(list(name="GS_A", desc=NULL, genes=c("MAPK14", "JAK1", "EGFR")),
    list(name="GS_B", desc="gene set B", genes=c("ABCA1", "DDR1", "DDR2")),
    list(name="GS_C", desc="gene set C", genes=NULL))
  testGmt <- BioQC::GmtList(testList)
  insertGmtListToGeMS(testGmt, geneFormat=0, source="Test", xref="PMID:000000")
  removeFromGeMS(setName=c("GS_A", "GS_B", "GS_C"), source="Test")
}

}
\seealso{
\code{\link{insertGmtListToGeMS}}
}
