% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\name{removeFromGeMSBody}
\alias{removeFromGeMSBody}
\title{Message body to remove one or gene sets of the same source and user from GeMS}
\usage{
removeFromGeMSBody(setName = "", source = "",
  user = ribiosUtils::whoami(), subtype = "")
}
\arguments{
\item{setName}{A vector of character strings, defining set names to be renamed. They must all have the same \code{source}, \code{user}, and \code{subtype}}

\item{source}{Character string, source of the gene set(s)}

\item{user}{Character string, user name}

\item{subtype}{Character string, subtype of the gene set(s)}
}
\value{
A list of genesets to be removed, to be sent as message body
}
\description{
Message body to remove one or gene sets of the same source and user from GeMS
}
\examples{
ribiosGSEA:::removeFromGeMSBody(setName=c("GS_A", "GS_B", "GS_C"), source="Test")
}
