% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ribiosGSEA-package.R
\docType{package}
\name{ribiosGSEA-package}
\alias{ribiosGSEA}
\alias{ribiosGSEA-package}
\title{ribiosGSEA: Gene-Set Enrichment Analysis Tools in 'ribios'}
\description{
Provides data structure and functions for gene-set analysis and post-processing of analysis results.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bedapub/ribiosGSEA}
  \item Report bugs at \url{https://github.com/bedapub/ribiosGSEA/issues}
}

}
\author{
\strong{Maintainer}: Jitao David Zhang \email{jitao_david.zhang@roche.com} (\href{https://orcid.org/0000-0002-3085-0909}{ORCID})

Other contributors:
\itemize{
  \item Balasz Banfai [contributor]
  \item F.Hoffmann-La Roche AG [copyright holder]
}

}
\keyword{internal}
