% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ronetGeneSetNamespace}
\alias{ronetGeneSetNamespace}
\title{Extract gene-set namespace from RONET GMT files}
\usage{
ronetGeneSetNamespace(gmtList)
}
\arguments{
\item{GmtList}{A GmtList object read from a RONET GMT file}
}
\value{
Character vector of the same length, indicating categories
}
\description{
Extract gene-set namespace from RONET GMT files
}
