% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,AnnoBroadGseaRes-method}
\alias{show,AnnoBroadGseaRes-method}
\title{Show a anonBroadGseaRes object}
\usage{
\S4method{show}{AnnoBroadGseaRes}(object)
}
\arguments{
\item{object}{A AnnoBroadGseaRes object 
export}
}
\description{
Show a anonBroadGseaRes object
}
