% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,AnnoBroadGseaResItem-method}
\alias{show,AnnoBroadGseaResItem-method}
\title{Show an AnnoBroadGseaResItem object}
\usage{
\S4method{show}{AnnoBroadGseaResItem}(object)
}
\arguments{
\item{object}{An annoBroadGseaResItem object 
export}
}
\description{
Show an AnnoBroadGseaResItem object
}
