% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,broadGseaResItem-method}
\alias{show,broadGseaResItem-method}
\title{Show a broadGseaResItem object}
\usage{
\S4method{show}{broadGseaResItem}(object)
}
\arguments{
\item{object}{A broadGseaResItem object 
export}
}
\description{
Show a broadGseaResItem object
}
