% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{sigGeneSet}
\alias{sigGeneSet}
\title{Return names of gene-sets that are significantly enriched given the 
FDR threshold}
\usage{
sigGeneSet(object, fdr)
}
\arguments{
\item{object}{A FisherResultList object}

\item{fdr}{Numeric, FDR value threshold}
}
\value{
A character vector
}
\description{
Return names of gene-sets that are significantly enriched given the 
FDR threshold
}
