% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{sigGeneSetTable}
\alias{sigGeneSetTable}
\title{Return a data.frame of significantly enriched gene-sets}
\usage{
sigGeneSetTable(object, fdr)
}
\arguments{
\item{object}{A FisherResultList object}

\item{fdr}{Numeric, FDR value threshold}
}
\value{
A \code{data.frame}
}
\description{
Return a data.frame of significantly enriched gene-sets
}
