% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[,AnnoBroadGseaRes,ANY,ANY,ANY-method}
\alias{[,AnnoBroadGseaRes,ANY,ANY,ANY-method}
\title{Subset an AnnoBroadGseaRes object}
\usage{
\S4method{[}{AnnoBroadGseaRes,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An AnnoBroadGseaRes object}

\item{i}{An integer or logical subsetting index}

\item{j}{Not used}

\item{...}{Not used}

\item{drop}{Not used}
}
\value{
A subset of the original data as an AnnoBroadGseaRes object
}
\description{
Subset an AnnoBroadGseaRes object
}
