% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[,FisherResultList,ANY,missing,missing-method}
\alias{[,FisherResultList,ANY,missing,missing-method}
\title{Subset a FisherResultList object by indexing}
\usage{
\S4method{[}{FisherResultList,ANY,missing,missing}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{A FisherResultList object}

\item{i}{An integer or logical subsetting index}

\item{j}{Not used}

\item{...}{Not used}

\item{drop}{Not used}
}
\value{
A subset of the original data as an FisherResultList object
}
\description{
Subset a FisherResultList object by indexing
}
