% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[,FisherResultList,character,character,missing-method}
\alias{[,FisherResultList,character,character,missing-method}
\title{Subset a FisherResultList object by namespace and name}
\usage{
\S4method{[}{FisherResultList,character,character,missing}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{A FisherResultList object}

\item{i}{Character string, gene-set namespace}

\item{j}{Character string, gene-set name}

\item{...}{Not used}

\item{drop}{Not used}
}
\description{
Subset a FisherResultList object by namespace and name
}
