% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{topGeneSetTable}
\alias{topGeneSetTable}
\title{Return a data.frame of top gene-sets with the lowest p-values}
\usage{
topGeneSetTable(object, N)
}
\arguments{
\item{object}{An FisherResultList object}

\item{N}{Integer, the number of returned gene-sets}
}
\description{
Return a data.frame of top gene-sets with the lowest p-values
}
