% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{topOrSigGeneSetTable}
\alias{topOrSigGeneSetTable}
\title{Return a data.frame of significantly enriched gene-sets with a minimum number}
\usage{
topOrSigGeneSetTable(object, fdr = 0.05, N = 10)
}
\arguments{
\item{object}{An FisherResultList object}

\item{fdr}{Numeric, the treshold of FDR value}

\item{N}{Integer, the number of returned gene-sets
The total number of returned gene-sets are determined by the maximum of
\code{N} and the counts of gene-sets that have \code{FDR} lower than
\code{fdr}.}
}
\value{
A \code{data.frame}.
}
\description{
Return a data.frame of significantly enriched gene-sets with a minimum number
}
