% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmt.R
\name{writeGmt}
\alias{writeGmt}
\title{Write an GmtList object into a file}
\usage{
writeGmt(gmtList, file)
}
\arguments{
\item{gmtList}{A \code{GmtList} object}

\item{file}{Character string, output file name}
}
\description{
Write an GmtList object into a file
}
\note{
The function will be moved to BioQC once the ribiosIO is reposited in CRAn
}
\examples{
gmtFile <- system.file("extdata", "example.gmt", package="ribiosGSEA")
mySet <- readGmt(gmtFile)[1:5]
myTempFile <- tempfile()
writeGmt(mySet, file=myTempFile)
readLines(myTempFile)
}
