% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{annoBroadGseaRes}
\alias{annoBroadGseaRes}
\title{Return annotated BROAD GSEA result}
\usage{
annoBroadGseaRes(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return annotated BROAD GSEA result
}
