% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{annoBroadGseaRes}
\alias{annoBroadGseaRes}
\title{Convert a list of annoBroadGseaResItem objects to a list}
\usage{
annoBroadGseaRes(object)
}
\arguments{
\item{object}{A list of annoBroadGseaResItem}
}
\value{
An \code{annoBroadGseaRes} object
}
\description{
Convert a list of annoBroadGseaResItem objects to a list
}
