% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnoBroadGseaResItem.R
\name{AnnoBroadGseaResItem}
\alias{AnnoBroadGseaResItem}
\title{Convert a BroadGseaResItem object to an AnnoBroadGseaResItem object}
\usage{
AnnoBroadGseaResItem(object, genes, geneValues)
}
\arguments{
\item{object}{A BroadGseaResItem object}

\item{genes}{A character string vector}

\item{geneValues}{A numeric vector}
}
\value{
An annoBroadGseaResItem object
}
\description{
Convert a BroadGseaResItem object to an AnnoBroadGseaResItem object
}
