% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{annoBroadGseaResItem}
\alias{annoBroadGseaResItem}
\title{Return annotated BROAD GSEA result item}
\usage{
annoBroadGseaResItem(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return annotated BROAD GSEA result item
}
