% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gse.R
\name{camera.EdgeResult}
\alias{camera.EdgeResult}
\title{Run CAMERA method using EdgeResult}
\usage{
\method{camera}{EdgeResult}(y, gmtList, doParallel = FALSE, ...)
}
\arguments{
\item{y}{A EdgeResult object}

\item{gmtList}{Gene set collections, for example read by
\code{\link[BioQC]{readGmt}}, with namespace.}

\item{doParallel}{Logical, whether \code{parallel::mclapply} should be used. Since at the current setting it makes a job running forever, use \code{TRUE} only if you are debugging the code.}

\item{...}{Not used

Note that the EdgeResult object must have a column 'GeneSymbol' in its
\code{fData}.}
}
\value{
A \code{data.frame} containing CAMERA results.
}
\description{
Run CAMERA method using EdgeResult
}
\examples{
exMat <- matrix(rpois(120, 10), nrow=20, ncol=6)
exGroups <- gl(2,3, labels=c("Group1", "Group2"))
exDesign <- model.matrix(~0+exGroups)
colnames(exDesign) <- levels(exGroups)
exContrast <- matrix(c(-1,1,1,-1), ncol=2, dimnames=list(c("Group1", "Group2"), 
  c("Group2.vs.Group1", "Group1.vs.Group2")))
exDescon <- DesignContrast(exDesign, exContrast, groups=exGroups)
exFdata <- data.frame(GeneSymbol=sprintf("GeneSymbol\%d", 1:nrow(exMat)))
exPdata <- data.frame(Name=sprintf("Sample\%d", 1:ncol(exMat)),
                     Group=exGroups)
exDgeList <- DGEList(exMat, genes=exFdata, samples=exPdata)
exDgeList <- edgeR::estimateDisp(exDgeList, exDesign)
exEdgeObject <- EdgeObject(exDgeList, exDescon)
exEdgeRes <- ribiosNGS::dgeWithEdgeR(exEdgeObject)
exGmt <- BioQC::GmtList(list(GeneSet1=sprintf("GeneSymbol\%d", 1:5),
  GeneSet2=sprintf("GeneSymbol\%d", 6:10)))
  
exCameraRes <- camera(exEdgeRes, exGmt)

}
