% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\docType{methods}
\name{fisherTest,character,GmtList,character-method}
\alias{fisherTest,character,GmtList,character-method}
\title{Perform Fisher's exact test on a GmtList object}
\usage{
\S4method{fisherTest}{character,GmtList,character}(genes, genesets,
  universe, gsCategory, makeUniqueNonNA = TRUE, checkUniverse = TRUE,
  useEASE = FALSE)
}
\arguments{
\item{genes}{character strings of gene list to be tested}

\item{genesets}{An GmtList object}

\item{universe}{Universe (background) gene list}

\item{makeUniqueNonNA}{Logical, whether genes and universe should be filtered to remove NA and made unique. The default is set to \code{TRUE}. When the uniqueness and absence of NA is ensured, this flag can be set to \code{FALSE} to accelerate the operation.}

\item{checkUniverse}{Logical, if \code{TRUE}, then genes that are in \code{genes} but are not in \code{universe} are appended to \code{universe}}

\item{useEASE}{Logical, whether to use the EASE method to report the p-value.}
}
\value{
A data.frame
}
\description{
Perform Fisher's exact test on a GmtList object
}
