% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{fisherTest}
\alias{fisherTest}
\title{Perform Fisher's exact test}
\usage{
fisherTest(genes, genesets, universe, ...)
}
\arguments{
\item{genes}{Genes}

\item{genesets}{Gene-sets}

\item{universe}{The universe of genes}

\item{...}{Other parameters}
}
\description{
Perform Fisher's exact test
}
