% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeMS.R
\docType{data}
\name{GeMS_BASE_URL}
\alias{GeMS_BASE_URL}
\title{GeMS base URL
To set GeMS base URL in your environment, use `GeMS_BASE_URL=value` in your
"~/.Renviron" file}
\format{
An object of class \code{character} of length 1.
}
\usage{
GeMS_BASE_URL
}
\description{
GeMS base URL
To set GeMS base URL in your environment, use `GeMS_BASE_URL=value` in your
"~/.Renviron" file
}
\keyword{datasets}
