% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gseaCoreEnrichGenes}
\alias{gseaCoreEnrichGenes}
\title{Return GSEA core enrichment genes}
\usage{
gseaCoreEnrichGenes(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA core enrichment genes
}
