% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaCoreEnrichThr}
\alias{gseaCoreEnrichThr}
\alias{gseaCoreEnrichThr,broadGseaResItem-method}
\alias{gseaCoreEnrichThr,annoBroadGseaRes-method}
\title{Return GSEA core enrichment score threshold}
\usage{
gseaCoreEnrichThr(object)

\S4method{gseaCoreEnrichThr}{broadGseaResItem}(object)

\S4method{gseaCoreEnrichThr}{annoBroadGseaRes}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA core enrichment score threshold
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get the threshold value of GSEA core enrichment
from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get the threshold value of GSEA core enrichment
from an annoBroadGseaRes object
}}

