% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaCoreEnrichThr}
\alias{gseaCoreEnrichThr}
\alias{gseaCoreEnrichThr,BroadGseaResItem-method}
\alias{gseaCoreEnrichThr,AnnoBroadGseaRes-method}
\title{Return GSEA core enrichment score threshold}
\usage{
gseaCoreEnrichThr(object)

\S4method{gseaCoreEnrichThr}{BroadGseaResItem}(object)

\S4method{gseaCoreEnrichThr}{AnnoBroadGseaRes}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA core enrichment score threshold
}
\section{Methods (by class)}{
\itemize{
\item \code{BroadGseaResItem}: Get the threshold value of GSEA core enrichment
from a BroadGseaResItem object

\item \code{AnnoBroadGseaRes}: Get the threshold value of GSEA core enrichment
from an AnnoBroadGseaRes object
}}

