% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaES}
\alias{gseaES}
\alias{gseaES,broadGseaResItem-method}
\alias{gseaES,annoBroadGseaRes-method}
\alias{gseaES,annoBroadGseaResList-method}
\title{Return GSEA enrichment scores}
\usage{
gseaES(object)

\S4method{gseaES}{broadGseaResItem}(object)

\S4method{gseaES}{annoBroadGseaRes}(object)

\S4method{gseaES}{annoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA enrichment scores
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get GSEA enrichment score from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get GSEA enrichment score from an annoBroadGseaRes object

\item \code{annoBroadGseaResList}: Get GSEA enrichment score from an annoBroadGseaResList object
}}

