% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaES}
\alias{gseaES}
\alias{gseaES,BroadGseaResItem-method}
\alias{gseaES,AnnoBroadGseaRes-method}
\alias{gseaES,AnnoBroadGseaResList-method}
\title{Return GSEA enrichment scores}
\usage{
gseaES(object)

\S4method{gseaES}{BroadGseaResItem}(object)

\S4method{gseaES}{AnnoBroadGseaRes}(object)

\S4method{gseaES}{AnnoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA enrichment scores
}
\section{Methods (by class)}{
\itemize{
\item \code{gseaES(BroadGseaResItem)}: Get GSEA enrichment score from a BroadGseaResItem object

\item \code{gseaES(AnnoBroadGseaRes)}: Get GSEA enrichment score from an AnnoBroadGseaRes object

\item \code{gseaES(AnnoBroadGseaResList)}: Get GSEA enrichment score from an AnnoBroadGseaResList object

}}
