% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gseaESprofile}
\alias{gseaESprofile}
\title{Return GSEA enrichment score profile}
\usage{
gseaESprofile(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA enrichment score profile
}
