% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaESprofile}
\alias{gseaESprofile}
\alias{gseaESprofile,BroadGseaResItem-method}
\title{Return GSEA enrichment score profile}
\usage{
gseaESprofile(object)

\S4method{gseaESprofile}{BroadGseaResItem}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA enrichment score profile
}
\section{Methods (by class)}{
\itemize{
\item \code{gseaESprofile(BroadGseaResItem)}: Get GSEA enrichment profile
from a BroadGseaResItem object

}}
