% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaFDR}
\alias{gseaFDR}
\alias{gseaFDR,BroadGseaResItem-method}
\alias{gseaFDR,AnnoBroadGseaRes-method}
\alias{gseaFDR,AnnoBroadGseaResList-method}
\title{Return GSEA FDR}
\usage{
gseaFDR(object)

\S4method{gseaFDR}{BroadGseaResItem}(object)

\S4method{gseaFDR}{AnnoBroadGseaRes}(object)

\S4method{gseaFDR}{AnnoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA FDR
}
\section{Methods (by class)}{
\itemize{
\item \code{BroadGseaResItem}: Get GSEA FDR values
from a BroadGseaResItem object

\item \code{AnnoBroadGseaRes}: Get GSEA FDR values
from an AnnoBroadGseaRes object

\item \code{AnnoBroadGseaResList}: Get GSEA FDR values
from an AnnoBroadGseaResList object
}}

