% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaFWER}
\alias{gseaFWER}
\alias{gseaFWER,broadGseaResItem-method}
\alias{gseaFWER,annoBroadGseaRes-method}
\alias{gseaFWER,annoBroadGseaResList-method}
\title{Return GSEA FWER values}
\usage{
gseaFWER(object)

\S4method{gseaFWER}{broadGseaResItem}(object)

\S4method{gseaFWER}{annoBroadGseaRes}(object)

\S4method{gseaFWER}{annoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA FWER values
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get GSEA FWER values
from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get GSEA FWER values
from an annoBroadGseaRes object

\item \code{annoBroadGseaResList}: Get GSEA FWER values
from an annoBroadGseaResList object
}}

