% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaFWER}
\alias{gseaFWER}
\alias{gseaFWER,BroadGseaResItem-method}
\alias{gseaFWER,AnnoBroadGseaRes-method}
\alias{gseaFWER,AnnoBroadGseaResList-method}
\title{Return GSEA FWER values}
\usage{
gseaFWER(object)

\S4method{gseaFWER}{BroadGseaResItem}(object)

\S4method{gseaFWER}{AnnoBroadGseaRes}(object)

\S4method{gseaFWER}{AnnoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA FWER values
}
\section{Methods (by class)}{
\itemize{
\item \code{BroadGseaResItem}: Get GSEA FWER values
from a BroadGseaResItem object

\item \code{AnnoBroadGseaRes}: Get GSEA FWER values
from an AnnoBroadGseaRes object

\item \code{AnnoBroadGseaResList}: Get GSEA FWER values
from an AnnoBroadGseaResList object
}}

