% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaNES}
\alias{gseaNES}
\alias{gseaNES,BroadGseaResItem-method}
\alias{gseaNES,AnnoBroadGseaRes-method}
\alias{gseaNES,AnnoBroadGseaResList-method}
\title{Return GSEA normalized enrichment scores}
\usage{
gseaNES(object)

\S4method{gseaNES}{BroadGseaResItem}(object)

\S4method{gseaNES}{AnnoBroadGseaRes}(object)

\S4method{gseaNES}{AnnoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA normalized enrichment scores
}
\section{Methods (by class)}{
\itemize{
\item \code{gseaNES(BroadGseaResItem)}: Get GSEA normalized enrichment score 
from a BroadGseaResItem object

\item \code{gseaNES(AnnoBroadGseaRes)}: Get GSEA normalized enrichment score 
from an AnnoBroadGseaRes object

\item \code{gseaNES(AnnoBroadGseaResList)}: Get GSEA normalized enrichment score 
from an AnnoBroadGseaResList object

}}
