% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gseaNP}
\alias{gseaNP}
\alias{gseaNP,BroadGseaResItem-method}
\alias{gseaNP,AnnoBroadGseaRes-method}
\alias{gseaNP,AnnoBroadGseaResList-method}
\title{Return GSEA number of permutation}
\usage{
gseaNP(object)

\S4method{gseaNP}{BroadGseaResItem}(object)

\S4method{gseaNP}{AnnoBroadGseaRes}(object)

\S4method{gseaNP}{AnnoBroadGseaResList}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return GSEA number of permutation
}
\section{Methods (by class)}{
\itemize{
\item \code{gseaNP(BroadGseaResItem)}: Get GSEA number of permutations
from a BroadGseaResItem object

\item \code{gseaNP(AnnoBroadGseaRes)}: Get GSEA number of permutations
from an AnnoBroadGseaRes object

\item \code{gseaNP(AnnoBroadGseaResList)}: Get GSEA number of permutations
from an AnnoBroadGseaResList object

}}
