% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{gsEffSize}
\alias{gsEffSize}
\title{Return gene-set effect size}
\usage{
gsEffSize(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set effect size
}
