% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsEffSize}
\alias{gsEffSize}
\alias{gsEffSize,FisherResult-method}
\alias{gsEffSize,FisherResultList-method}
\title{Return the effective size of gene-set(s)}
\usage{
gsEffSize(object, ...)

\S4method{gsEffSize}{FisherResult}(object)

\S4method{gsEffSize}{FisherResultList}(object)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\value{



}
\description{
Return the effective size of gene-set(s)
}
\section{Methods (by class)}{
\itemize{
\item \code{FisherResult}: Effective sizes of gene-set, returning an integer value.

\item \code{FisherResultList}: Effective sizes of Gene-sets, returning an integer vector.
}}

