% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGenes<-}
\alias{gsGenes<-}
\alias{gsGenes<-,AnnoBroadGseaResItem,character-method}
\title{gsGenes-set}
\usage{
gsGenes(object) <- value

\S4method{gsGenes}{AnnoBroadGseaResItem,character}(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Value}
}
\description{
Set gene-set genes
}
\section{Functions}{
\itemize{
\item \code{gsGenes<-,AnnoBroadGseaResItem,character-method}: Assign gene-set genes to AnnoBroadGseaResItem
}}

