% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGeneValues<-}
\alias{gsGeneValues<-}
\alias{gsGeneValues<-,annoBroadGseaResItem,numeric-method}
\title{gsGeneValues-set}
\usage{
gsGeneValues(object) <- value

\S4method{gsGeneValues}{annoBroadGseaResItem,numeric}(object) <- value
}
\arguments{
\item{object}{An object}

\item{value}{Value}
}
\description{
Set gene-set gene statistics (values)
}
\section{Functions}{
\itemize{
\item \code{gsGeneValues<-,annoBroadGseaResItem,numeric-method}: Assign values associated with gene-set genes to
an annoBraoadGseaResItem object
}}

