% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsGeneValues}
\alias{gsGeneValues}
\alias{gsGeneValues,annoBroadGseaResItem-method}
\alias{gsGeneValues,annoBroadGseaRes-method}
\title{Return gene-set gene values}
\usage{
gsGeneValues(object)

\S4method{gsGeneValues}{annoBroadGseaResItem}(object)

\S4method{gsGeneValues}{annoBroadGseaRes}(object)
}
\arguments{
\item{object}{An object}
}
\description{
Return gene-set gene values
}
\section{Methods (by class)}{
\itemize{
\item \code{annoBroadGseaResItem}: Return values associated with the genes in a
gene-set in an annoBroadGseaResItem object in a numeric vector.

\item \code{annoBroadGseaRes}: Return values associated with the genes in a
gene-set in an annoBroadGseaRes object in a list of numeric vectors.
}}

