% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/AllMethods.R
\name{gsName}
\alias{gsName}
\alias{gsName,broadGseaResItem-method}
\alias{gsName,annoBroadGseaRes-method}
\alias{gsName,FisherResult-method}
\alias{gsName,FisherResultList-method}
\alias{gsName,GmtList-method}
\title{Return gene-set name}
\usage{
gsName(object, ...)

\S4method{gsName}{broadGseaResItem}(object)

\S4method{gsName}{annoBroadGseaRes}(object)

\S4method{gsName}{FisherResult}(object)

\S4method{gsName}{FisherResultList}(object, ...)

\S4method{gsName}{GmtList}(object)

\S4method{gsName}{FisherResultList}(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{Other parameters}
}
\description{
Return gene-set name
}
\section{Methods (by class)}{
\itemize{
\item \code{broadGseaResItem}: Get gene-set name from a broadGseaResItem object

\item \code{annoBroadGseaRes}: Get gene-set name from an annoBroadGseaRes object

\item \code{FisherResult}: Get gene-set name from a FisherResult object

\item \code{FisherResultList}: Get gene-set name from a FisherResultList object

\item \code{GmtList}: Get gene-set name from a GmtList object

\item \code{FisherResultList}: Get gene-set name from a FisherResultList object
}}

