% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{isGseaCoreEnrich}
\alias{isGseaCoreEnrich}
\title{Return a vector of logical values, indicating whether genes belong to core
enrichment or not}
\usage{
isGseaCoreEnrich(object)
}
\arguments{
\item{object}{An \code{annoBroadGseaResItem} object}
}
\value{
A logical vector
}
\description{
Return a vector of logical values, indicating whether genes belong to core
enrichment or not
}
