% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{isSigGeneSet}
\alias{isSigGeneSet}
\title{Return a logical vector indicating whether a gene-set is significantly
enriched or not, given the FDR threshold}
\usage{
isSigGeneSet(object, fdr = 0.05)
}
\arguments{
\item{object}{A FisherResultList object}

\item{fdr}{Numeric, FDR value threshold}
}
\value{
A logical vector
}
\description{
Return a logical vector indicating whether a gene-set is significantly
enriched or not, given the FDR threshold
}
