% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher.R
\name{minPvalue}
\alias{minPvalue}
\title{Return the minimal p-value from a FisherResultList}
\usage{
minPvalue(object)
}
\arguments{
\item{object}{A FisherResultList object}
}
\value{
A numeric value
}
\description{
Return the minimal p-value from a FisherResultList
}
