% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,annoBroadGseaRes-method}
\alias{show,annoBroadGseaRes-method}
\title{Show a anonBroadGseaRes object}
\usage{
\S4method{show}{annoBroadGseaRes}(object)
}
\arguments{
\item{object}{A annoBroadGseaRes object 
export}
}
\description{
Show a anonBroadGseaRes object
}
