% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,annoBroadGseaResItem-method}
\alias{show,annoBroadGseaResItem-method}
\title{Show an annoBroadGseaResItem object}
\usage{
\S4method{show}{annoBroadGseaResItem}(object)
}
\arguments{
\item{object}{An annobroadGseaResItem object 
export}
}
\description{
Show an annoBroadGseaResItem object
}
