% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{show,BroadGseaResItem-method}
\alias{show,BroadGseaResItem-method}
\title{Show a BroadGseaResItem object}
\usage{
\S4method{show}{BroadGseaResItem}(object)
}
\arguments{
\item{object}{A BroadGseaResItem object 
export}
}
\description{
Show a BroadGseaResItem object
}
