% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sigGeneSet}
\alias{sigGeneSet}
\title{Return a character vector indicating significantly enriched gene-sets}
\usage{
sigGeneSet(object, fdr, ...)
}
\arguments{
\item{object}{An object}

\item{fdr}{Numeric, FDR threshold}
}
\description{
Return a character vector indicating significantly enriched gene-sets
}
