% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{sigGeneSetTable}
\alias{sigGeneSetTable}
\title{Return a table of significantly enriched gene-ests}
\usage{
sigGeneSetTable(object, fdr, ...)
}
\arguments{
\item{object}{An object}

\item{fdr}{Numeric, FDR threshold}

\item{...}{Other parameters}
}
\description{
Return a table of significantly enriched gene-ests
}
