% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllMethods.R
\name{[,annoBroadGseaRes,ANY,ANY,ANY-method}
\alias{[,annoBroadGseaRes,ANY,ANY,ANY-method}
\title{Subset an annoBroadGseaRes object}
\usage{
\S4method{[}{annoBroadGseaRes,ANY,ANY,ANY}(x, i, j, ..., drop = FALSE)
}
\arguments{
\item{x}{An annoBroadGseaRes object}

\item{i}{An integer or logical subsetting index}

\item{j}{Not used}

\item{...}{Not used}

\item{drop}{Not used}
}
\value{
A subset of the original data as an annoBroadGseaRes object
}
\description{
Subset an annoBroadGseaRes object
}
