% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{topGeneSetTable}
\alias{topGeneSetTable}
\title{Return top enriched gene-sets}
\usage{
topGeneSetTable(object, N, ...)
}
\arguments{
\item{object}{An object}

\item{N}{integer, number of top gene-sets to be returned}
}
\description{
Return top enriched gene-sets
}
