% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R
\name{topOrSigGeneSetTable}
\alias{topOrSigGeneSetTable}
\title{Return top or significantly enriched gene-sets}
\usage{
topOrSigGeneSetTable(object, N, fdr, ...)
}
\arguments{
\item{object}{An object}

\item{N}{integer, number of top gene-sets to be returned}

\item{fdr}{Numeric, FDR threshold}

\item{...}{Other parameters}
}
\description{
Return top or significantly enriched gene-sets
}
